net/virtio-user: fix not working on 32-bit system
virtio-user cannot work on 32-bit system as higher 32-bit of the
addr field (64-bit) in the desc is filled with non-zero value
which should not happen for a 32-bit system.

This is a regression bug. For 32-bit system, the first 4 bytes
is the virtual address, with following 8 bytes pointing to
physical addr. With below wrong definition, both virtual address
and lower 4 bytes of physical addr are obtained.
  #define VIRTIO_MBUF_ADDR(mb, vq) \
	(*(uint64_t *)((uintptr_t)(mb) + (vq)->offset))

Fixes: 25f80d108780 ("net/virtio: fix packet corruption")
Cc: stable@dpdk.org

Signed-off-by: Jianfeng Tan <jianfeng.tan@intel.com>