rte_flow: add attribute for signature match
Add new attribute "sig_match" to rte_flow_attr.
This attribute indicate if current flow take "perfect match"
or "signature match".
With perfect match (by default), if a packet does not match pattern,
actions will not be taken. (this is identical with current behavior )
With signature match, if a packet does not match pattern, it still
has the possibility to trigger the actions, this happens when device
think the signature of the pattern is matched.
Signature match is expected to have better performance than perfect
match, but the cost is accuracy.
When a flow rule with this attribute set, identical behavior can ONLY
be guaranteed if packet matches the pattern, since different device
may have different implementation of signature calculation algorithm.

Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>