Correctly handle malloc_elem resize with padding
Currently when a malloc_elem is split after resizing, any padding
present in the elem is ignored.  This causes the resized elem to be too
small when padding is present, and user data can overwrite the beginning
of the following malloc_elem.

Solve this by including the size of the padding when computing where to
split the malloc_elem.

Signed-off-by: Jamie Lavigne <lavignen@amazon.com>