lib/gro: add TCP/IPv4 GRO support
In this patch, we introduce six APIs to support TCP/IPv4 GRO.
- gro_tcp_tbl_create: create a TCP reassembly table, which is used to
    merge packets.
- gro_tcp_tbl_destroy: free memory space of a TCP reassembly table.
- gro_tcp_tbl_flush: flush packets in the TCP reassembly table.
- gro_tcp_tbl_timeout_flush: flush timeout packets in the TCP
    reassembly table.
- gro_tcp4_reassemble: merge an inputted packet.
- gro_tcp4_tbl_cksum_update: update TCP and IPv4 header checksums for
    all merged packets in the TCP reassembly table.

In TCP GRO, we use a table structure, called TCP reassembly table, to
reassemble packets. Both TCP/IPv4 and TCP/IPv6 GRO use the same table
structure. A TCP reassembly table includes a flow array and a item array,
where the flow array is used to record flow information and the item
array is used to record packets information.

Each element in the flow array records the information of one flow,
which includes two parts:
- key: the criteria of the same flow. If packets have the same key
    value, they belong to the same flow.
- start_index: the index of the first incoming packet of this flow in
    the item array. With start_index, we can locate the first incoming
    packet of this flow.
Each element in the item array records one packet information. It mainly
includes two parts:
- pkt: packet address
- next_pkt_index: index of the next packet of the same flow in the item
    array. All packets of the same flow are chained by next_pkt_index.
    With next_pkt_index, we can locate all packets of the same flow
    one by one.

To process an incoming packet, we need three steps:
a. check if the packet should be processed. Packets with the following
    properties won't be processed:
	- packets without data;
	- packets with wrong checksums;
	- fragmented packets.
b. traverse the flow array to find a flow which the packet belongs to.
    If not find, insert a new flow and store the packet into the item
    array.
c. locate the first packet of this flow in the item array via
    start_index. Then traverse all packets of this flow one by one via
    next_pkt_index. If find one packet to merge with the incoming packet,
    merge them but without updating checksums. If not, allocate one item
    in the item array to store the incoming packet and update
    next_pkt_index value.

For better performance, we don't udpate header checksums once two
packets are merged. The header checksums are updated only when packets
are flushed from TCP reassembly tables.

Signed-off-by: Jiayu Hu <jiayu.hu@intel.com>