net/bond: calculate number of bonding tx/rx queues
Fixes: 2efb58cb ("bond: new link bonding library")

This patch fixes the maximum number of tx an rx queues supported by a
bonding device return by the rte_eth_dev_info_get function. The bonding
device now calculates the maximum number of supported tx and rx queues
based on the slaves bound to the bonded device, with the minimum values
of tx and rx queues from the device slaves being the bonded devices
maximum, as each slave must be able to support the same number of tx
and rx queues.

Signed-off-by: Declan Doherty <declan.doherty@intel.com>