usertools: add dpdk run script
I use a script like this one with pktgen and wanted to see if DPDK
would be interested in this application.

The following script adds support for executing applications using
a configuration file. The configuration file is formatted as a
python data file to be loaded by the run.py script.

Inside the configuration 'default.cfg' is the command line arguments
needed to execute the application. Any number of configuration  files
are allows for a single application for different execution. Normally
the configuration file can be in the application or example directory
or placed in a RTE_SDK/cfg directory.

The configuration file contains information about setup of the system
and how to run the application. The run.py script can setup the system
and/or execute the application with a simple command.

./run.py -s default # for setup
./run.py default # to execute the application.

I am not a great Python coder and any suggestions/patches would be great.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>