eal: fix hotplug add / remove
New device should be represented by an rte_devargs prior to being
plugged.

Device parameters are available to rte_devices via their devargs field.
This field should be set up as soon as possible, as central information
are stored within, such as the device name which is used to search
the newly scanned device before plugging it in.

When a device is introduced using the hotplug API, it is implicitly
whitelisted. As such, it can conflict with existing bus configuration.
The new rte_devargs uses the new rte_eal_devargs_insert function that
supersedes previous rte_devargs, allowing to force the insertion of new
devices.

Upon device removal, the newly allocated rte_devargs are freed.

Fixes: a3ee360f4440 ("eal: add hotplug add/remove device")

Signed-off-by: Gaetan Rivet <gaetan.rivet@6wind.com>