net/mlx4: workaround to verbs wrong error return
Current mlx4 OFED version has bug which returns error to
ibv destroy functions when the device was plugged out, in
spite of the resources were destroyed correctly.

Hence, failsafe PMD was aborted, only in debug mode, when
it tries to remove the device in plug-out process.

The workaround removed the ibv destroy assertions.

DPDK 18.02 release should work with OFED-4.2 which will
include the verbs fix to this bug, then, this patch will
be removed.

Signed-off-by: Matan Azrad <matan@mellanox.com>
Cc: stable@dpdk.org