net/virtio-user: send kick to tx queue to notify backend on initialization
Acccording to the spec, https://fossies.org/linux/qemu/docs/specs/vhost-user.txt

client must start ring upon receiving a kick (that is, detecting that file
descriptor is reachable) on the descriptor specified by
VHOST_USER_SET_VRING_KICK.

The code sends a kick to the rx queue. It is missing sending a kick for the
tx queue. This patch is to add the missing code to comply with the spec.

Signed-off-by: Steven <sluong@cisco.com>